Split style files
